+++ src/main/java/team/blackjack/Application.java
@@ -1,7 +1,7 @@
 package team.blackjack;
 
 import team.blackjack.config.AppConfig;
-import team.blackjack.controler.BlackJackController;
+import team.blackjack.controller.BlackJackController;
 
 public class Application {
     public static void main(String[] args) {
